+++ packages/core/src/domain/entities/sensor.entity.ts
@@ -4,7 +4,7 @@ import { EntityConfig, CommandSchemaOrCEL } from './base.entity.js';
 import { StateSchema, StateNumSchemaOrCEL } from '../../protocol/types.js';
 
 export interface SensorEntity extends EntityConfig {
-  state: StateSchema;
+  state?: StateSchema;
   state_number?: StateNumSchemaOrCEL;
   command_update?: CommandSchemaOrCEL;
 }